Split config into focused modules
